Allow registration by project or group id
Registration by a specific token is the traditional way to register a
runner -- and indeed, the only way to register a shared runner.
(AFAICT)  However, they sometimes change, and it would be wicked
annoying to have all of your runner registrations invalidated because
someone hit "reset token" in your largest group.

There are now three ways to register a runner:

* "register_token"

Unchanged.

* "register_project_id" or "register_group_id"

Specifying either of these two options causes the provider to
interrogate GitLab to determine the correct token to be used; the
retrieved token is then used to register the runner.  The registration
token is not needed past this point, and is not stored in state.