Add `hashSum()` to hash the registration token
The only reason we need to keep a registration token around is to
determine if it has changed, and if we should then recreate the runner
registration.  However, we can compare a checksum of the token just as
readily, so let's do that.  (e.g. via `StateFunc`)

This function was liberated from `terraform-provider-aws`.